Fixed bug in example2. result should be uint16_t

+++ Examples/example2.c
@@ -9,7 +9,7 @@ uint16_t multiply(uint8_t a, uint8_t b) {
 
 int main(void) {
    uint8_t foo = 50, bar = 10;
-   uint8_t result;
+   uint16_t result;
    /* return val implicitly typecasted (bad) */
    result = multiply(foo, bar);
    printf("foo = % d bar = % d result = % d\n", foo, bar, result);
